Последний файл

+++ app/src/main/java/com/esd/esd/biathlontimer/Activities/ViewPagerActivity.java
@@ -114,6 +114,7 @@ public class ViewPagerActivity extends AppCompatActivity
 
     private Competition _currentCompetition;
 
+    private boolean _needDeleteTables = false;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -128,13 +129,14 @@ public class ViewPagerActivity extends AppCompatActivity
         String startType = intent.getStringExtra("CompetitionStartType");
         String interval = intent.getStringExtra("CompetitionInterval");
         String checkPoints = intent.getStringExtra("CompetitionCheckPointsCount");
+        _needDeleteTables = Boolean.valueOf(intent.getStringExtra("NeedDelete"));
         _currentCompetition = new Competition(name, date, this);
         _currentCompetition.SetCompetitionSettings(startType, interval, checkPoints, "");
         FindAllViews();
 
         // Создание диалогового окна
         _addDialogBuilder = new AlertDialog.Builder(this);
-        _addDialogBuilder.setTitle(TitleDialog);  // заголовок
+        _addDialogBuilder.setTitle(TitleDialog);
         _addDialogBuilder.setView(_dialogForm);
 
         // Действия по кнопке "Добавить"
@@ -142,7 +144,7 @@ public class ViewPagerActivity extends AppCompatActivity
             public void onClick(DialogInterface dialog, int arg1)
             {
                 Participant participant = new Participant(_numberDialog.getText().toString(),_nameDialog.getText().toString(),
-                        _countryDialog.getText().toString(), _birthdayDialog.getText().toString());
+                        _countryDialog.getText().toString(), _birthdayDialog.getText().toString(),"");
                 AddRowParticipantList(participant);
                 _acceptParticipantImBtn.setVisibility(View.VISIBLE);
                 if(_dbSaver.SaveParticipantToDatabase(participant, DatabaseProvider.DbParticipant.TABLE_NAME))
@@ -213,12 +215,12 @@ public class ViewPagerActivity extends AppCompatActivity
                     cellsCount = 3;
                 }
 
-                localParticipant = new Participant(number, name, country, year);
+                localParticipant = new Participant(number, name, country, year, "");
 
                 _dbSaver.DeleteParticipant(localParticipant, DatabaseProvider.DbParticipant.TABLE_NAME);
                 _dbSaver.DeleteParticipant(localParticipant, _currentCompetition.GetDbParticipantPath());
                 localParticipant = new Participant(_numberRenameDialog.getText().toString(), _nameRenameDialog.getText().toString(),
-                        _countryRenameDialog.getText().toString(), _birthdayRenameDialog.getText().toString());
+                        _countryRenameDialog.getText().toString(), _birthdayRenameDialog.getText().toString(), "");
                 _dbSaver.SaveParticipantToDatabase(localParticipant, DatabaseProvider.DbParticipant.TABLE_NAME);
                 if(currTable == _tableLayoutParticipantList)
                 {
@@ -575,6 +577,67 @@ public class ViewPagerActivity extends AppCompatActivity
 
     }
 
+    private void SortByYear(TableLayout table,boolean sortState, boolean sortByYear)
+    {
+        Participant[] localArr = GetParticipantsFromTable(table);
+        int participantCount = localArr.length;
+        table.removeAllViews();
+        int k = 0;
+        Participant helpPart;
+        int firstParticipantData;
+        int secondParticipantData;
+        while(k != participantCount)
+        {
+            for(int i = 0; i < participantCount - k - 1; i++)
+            {
+                if(sortByYear)
+                {
+                    firstParticipantData = Integer.valueOf(localArr[i].GetBirthYear());
+                    secondParticipantData = Integer.valueOf(localArr[i + 1].GetBirthYear());
+                }
+                else
+                {
+                    try {
+                        firstParticipantData = Integer.valueOf(localArr[i].GetNumber());
+                    }catch (Exception e) {firstParticipantData = 0;}
+                    try {
+                        secondParticipantData = Integer.valueOf(localArr[i + 1].GetNumber());
+                    }catch (Exception e1) {secondParticipantData = 0;}
+                }
+                if(sortState)
+                {
+                    if (firstParticipantData > secondParticipantData) {
+                        helpPart = localArr[i];
+                        localArr[i] = localArr[i + 1];
+                        localArr[i + 1] = helpPart;
+                    }
+                }
+                else
+                {
+                    if (firstParticipantData < secondParticipantData) {
+                        helpPart = localArr[i];
+                        localArr[i] = localArr[i + 1];
+                        localArr[i + 1] = helpPart;
+                    }
+                }
+            }
+            k++;
+        }
+        if(table == _tableLayoutDataBaseList)
+        {
+            for (int j = 0; j < participantCount; j++)
+            {
+                AddRowParticipantFromBase(localArr[j]);
+            }
+        }
+        else
+        {
+            for (int j = 0; j < participantCount; j++)
+            {
+                AddRowParticipantList(localArr[j]);
+            }
+        }
+    }
     // Обработка нажатий меню сортировок
     public void OnClickMenu(View view)
     {
@@ -600,14 +663,15 @@ public class ViewPagerActivity extends AppCompatActivity
                                     Toast.makeText(getApplicationContext(),"Сортировка списка участников по имени",Toast.LENGTH_SHORT).show();
                                     return true;
                                 case R.id.dataSort:
-                                    SortTableBy(_tableLayoutParticipantList, DatabaseProvider.DbParticipant.COLUMN_NAME, item.isChecked());
+                                    SortByYear(_tableLayoutParticipantList, item.isChecked(), true);
                                     Toast.makeText(getApplicationContext(),"Сортировка списка участников по дате",Toast.LENGTH_SHORT).show();
                                     return true;
                                 case R.id.countrySort:
-                                    SortTableBy(_tableLayoutParticipantList, DatabaseProvider.DbParticipant.COLUMN_NAME, item.isChecked());
+                                    SortTableBy(_tableLayoutParticipantList, DatabaseProvider.DbParticipant.COLUMN_COUNTRY, item.isChecked());
                                     Toast.makeText(getApplicationContext(),"Сортировка списка участников по региону",Toast.LENGTH_SHORT).show();
                                     return true;
                                 case R.id.numberSort:
+                                    SortByYear(_tableLayoutParticipantList, item.isChecked(), false);
                                     Toast.makeText(getApplicationContext(),"Сортировка списка участников по номеру",Toast.LENGTH_SHORT).show();
                                     return true;
                                 default:
@@ -638,7 +702,8 @@ public class ViewPagerActivity extends AppCompatActivity
                                     Toast.makeText(getApplicationContext(),"Сортировка базы данных по имени",Toast.LENGTH_SHORT).show();
                                     return true;
                                 case R.id.dataSort:
-                                    SortTableBy(_tableLayoutDataBaseList, DatabaseProvider.DbParticipant.COLUMN_YEAR, item.isChecked());
+                                    //SortTableBy(_tableLayoutDataBaseList, DatabaseProvider.DbParticipant.COLUMN_YEAR, item.isChecked());
+                                    SortByYear(_tableLayoutDataBaseList, item.isChecked(),true);
                                     Toast.makeText(getApplicationContext(),"Сортировка базы данных по дате",Toast.LENGTH_SHORT).show();
                                     return true;
                                 case R.id.countrySort:
@@ -657,18 +722,30 @@ public class ViewPagerActivity extends AppCompatActivity
         }
     }
 
-    private Participant[] GetParticipantsFromTable()
+    private Participant[] GetParticipantsFromTable(TableLayout table)
     {
-        int participantCount = _tableLayoutParticipantList.getChildCount();
+        int participantCount = table.getChildCount();
         Participant[] localArr = new Participant[participantCount];
-        String[] dataArr = new String[4];
+        String[] dataArr;
+        int cellsCount;
+        if(table == _tableLayoutDataBaseList)
+        {
+            cellsCount = 3;
+            dataArr = new String[cellsCount];
+        }
+        else
+        {
+            cellsCount = 4;
+            dataArr = new String[cellsCount];
+        }
         for(int i = 0; i<participantCount; i++)
         {
-            for(int j = 0; j < 4; j++)
+            for(int j = 0; j < cellsCount; j++)
             {
-                dataArr[j] = ((TextView)((TableRow) _tableLayoutParticipantList.getChildAt(i)).getChildAt(j)).getText().toString();
+                dataArr[j] = ((TextView)((TableRow) table.getChildAt(i)).getChildAt(j)).getText().toString();
             }
-            localArr[i] = new Participant(dataArr[0],dataArr[1], dataArr[3], dataArr[2]);
+            if(table == _tableLayoutParticipantList) localArr[i] = new Participant(dataArr[0],dataArr[1], dataArr[3], dataArr[2],"");
+            else localArr[i] = new Participant("",dataArr[0], dataArr[2], dataArr[1], "");
         }
 
         return localArr;
@@ -709,11 +786,11 @@ public class ViewPagerActivity extends AppCompatActivity
                 flag = false;
                 if(table == _tableLayoutParticipantList)
                 {
-                    localArr.add(new Participant(dataArr[0],dataArr[1], dataArr[3], dataArr[2]));
+                    localArr.add(new Participant(dataArr[0],dataArr[1], dataArr[3], dataArr[2],""));
                 }
                 else
                 {
-                    localArr.add(new Participant("",dataArr[0], dataArr[2], dataArr[1]));
+                    localArr.add(new Participant("",dataArr[0], dataArr[2], dataArr[1], ""));
                 }
                 if(needDelete) table.removeViewAt(i);
                 else
@@ -731,7 +808,7 @@ public class ViewPagerActivity extends AppCompatActivity
     {
         CompetitionSaver competitionSaver = new CompetitionSaver(this);
         competitionSaver.SaveCompetitionToDatabase(_currentCompetition);
-
+        _needDeleteTables = false;
         Toast.makeText(getApplicationContext(),"Сохранить список и перейти к соревнованию",Toast.LENGTH_SHORT).show();
     }
 
@@ -807,7 +884,7 @@ public class ViewPagerActivity extends AppCompatActivity
     public void OnClickAcceptDataBase(View view)
     {
         Participant[] localArr = GetCheckedParticipants(_tableLayoutDataBaseList, true);
-        Participant[] localCompParticipants = GetParticipantsFromTable();
+        Participant[] localCompParticipants = GetParticipantsFromTable(_tableLayoutParticipantList);
         boolean localFlag;
         for(int i = 0; i<localArr.length; i++)
         {
@@ -909,10 +986,15 @@ public class ViewPagerActivity extends AppCompatActivity
             dialog.setPositiveButton("Да", new DialogInterface.OnClickListener() {
                 @Override
                 public void onClick(DialogInterface dialog, int which) {
-                    DatabaseProvider provider = new DatabaseProvider(ViewPagerActivity.this);
-                    provider.DeleteTable(_currentCompetition.GetDbParticipantPath());
-                    provider.DeleteTable(_currentCompetition.GetSettingsPath());
+                    if(_needDeleteTables)
+                    {
+                        DatabaseProvider provider = new DatabaseProvider(ViewPagerActivity.this);
+                        provider.DeleteTable(_currentCompetition.GetDbParticipantPath());
+                        provider.DeleteTable(_currentCompetition.GetSettingsPath());
+                    }
+                    Intent intent = new Intent(ViewPagerActivity.this, MainActivity.class);
                     ViewPagerActivity.this.finish();
+                    startActivity(intent);
                 }
             });
             dialog.setNegativeButton("Нет", new DialogInterface.OnClickListener() {
